deliverer/requirements.txt
httpx
tenacity
python-dotenv
